+++ pidev-backend/src/main/java/tn/esprit/pidev/services/chat/IGestionMessage.java
@@ -6,13 +6,13 @@ import java.util.List;
 
 public interface IGestionMessage {
 
-    public Message sendMessage (Integer senderUser, Long discussion, String message);
+    public Message sendMessage (Long senderUser, Long discussion, String message);
 
     public Message modifyMessage (Long id, String message);
 
     public boolean deleteMessage (Long id);
 
-    public Message replyMessage(Integer userSender, Long discussion, Long message, String reply);
+    public Message replyMessage(Long userSender, Long discussion, Long message, String reply);
 
     public List<Message> retrieveAllMessages(Long id);
