+++ pidev-backend/src/main/java/tn/esprit/pidev/services/chat/IGestionReaction.java
@@ -4,7 +4,7 @@ import tn.esprit.pidev.entities.chat.Reaction;
 
 public interface IGestionReaction {
 
-    public Reaction addReaction (Integer user, Long message, String reaction);
+    public Reaction addReaction (Long user, Long message, String reaction);
 
     public boolean deleteReaction (Long id);
